Extract Prefix class out of ArgumentTokenizer class
The Prefix class is a nested class in ArgumentTokenizer.

It is unnecessarily tightly coupled with ArgumentTokenizer. The Prefix
class is used by other classes that do not need access to the
ArgumentTokenizer class.

Let's extract the Prefix class into its own class.
