+++ src/main/java/seedu/address/logic/parser/ArgumentTokenizer.java
@@ -156,40 +156,6 @@ public class ArgumentTokenizer {
     }
 
     /**
-     * A prefix that marks the beginning of an argument.
-     * e.g. 't/' in 'add James t/ friend'
-     */
-    public static class Prefix {
-        final String prefix;
-
-        Prefix(String prefix) {
-            this.prefix = prefix;
-        }
-
-        String getPrefix() {
-            return this.prefix;
-        }
-
-        @Override
-        public int hashCode() {
-            return this.prefix == null ? 0 : this.prefix.hashCode();
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (!(obj instanceof Prefix)) {
-                return false;
-            }
-            if (obj == this) {
-                return true;
-            }
-
-            Prefix otherPrefix = (Prefix) obj;
-            return otherPrefix.getPrefix().equals(getPrefix());
-        }
-    }
-
-    /**
      * Represents a prefix's position in an arguments string
      */
     private class PrefixPosition {
