+++ src/main/java/seedu/address/logic/parser/CliSyntax.java
@@ -1,7 +1,5 @@
 package seedu.address.logic.parser;
 
-import seedu.address.logic.parser.ArgumentTokenizer.Prefix;
-
 /**
  * Contains Command Line Interface (CLI) syntax definitions common to multiple commands
  */
