+++ src/test/java/seedu/address/logic/parser/ArgumentTokenizerTest.java
@@ -7,8 +7,6 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
-import seedu.address.logic.parser.ArgumentTokenizer.Prefix;
-
 public class ArgumentTokenizerTest {
 
     private final Prefix unknownPrefix = new Prefix("--u");
